chore: Moved Grid and Cell into their own files
