+++ src/GridRefactor/Cell/index.ts
@@ -1,0 +1,1 @@
+export { Cell } from './Cell.component';
