+++ src/GridRefactor/Grid/index.ts
@@ -1,0 +1,1 @@
+export { Grid } from './Grid.component';
