+++ src/GridRefactor/index.ts
@@ -1,1 +1,2 @@
-export * from './GridRefactor.component';
+export { Cell } from './Cell';
+export { Grid } from './Grid';
